Move `ThreadContextTestAccess` to `org.apache.logging.log4j.core.test` to eliminate split package
