+++ log4j-core-test/src/main/java/org/apache/logging/log4j/core/test/ThreadContextTestAccess.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.logging.log4j.core;
+package org.apache.logging.log4j.core.test;
 
 import static org.junit.jupiter.api.Assertions.fail;
