+++ src/changelog/.3.x.x/4315_threadcontexttestaccess_package.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<entry xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns="https://logging.apache.org/xml/ns"
+       xsi:schemaLocation="https://logging.apache.org/xml/ns https://logging.apache.org/xml/ns/log4j-changelog-0.xsd"
+       type="changed">
+  <issue id="4315" link="https://github.com/apache/logging-log4j2/pull/4315"/>
+  <description format="asciidoc">Move `ThreadContextTestAccess` from `org.apache.logging.log4j.core` to `org.apache.logging.log4j.core.test` to remove a split package.</description>
+</entry>
